+++ gemlib/n_vq_dflt_ctab.c
@@ -20,7 +20,7 @@
  */
 
 short
-vq_dflt_ctab (short handle, long ctab_length, COLOR_TAB *ctab)
+vq_dflt_ctab (short handle, int32_t ctab_length, COLOR_TAB *ctab)
 {
 	short vdi_control[VDI_CNTRLMAX];
