+++ gemlib/n_vs_hilite_color.c
@@ -20,7 +20,7 @@
  */
 
 short
-vs_hilite_color (short handle, long color_space, COLOR_ENTRY * hilite_color)
+vs_hilite_color (short handle, int32_t color_space, COLOR_ENTRY * hilite_color)
 {
 	short vdi_control[VDI_CNTRLMAX]; 
 	short vdi_intin[6];
