+++ gemlib/n_vs_max_color.c
@@ -20,7 +20,7 @@
  */
 
 short
-vs_max_color (short handle, long color_space, COLOR_ENTRY * min_color)
+vs_max_color (short handle, int32_t color_space, COLOR_ENTRY * min_color)
 {
 	short vdi_control[VDI_CNTRLMAX]; 
 	short vdi_intin[6];
